@@ -20,7 +20,7 @@
     }
   ],
   "require": {
-    "php": ">=5.5.9",
+    "php": ">=5.6",
     "guzzlehttp/guzzle": "^6.2.1 || ^7.0.1",
     "guzzlehttp/uri-template": "^0",
     "phpseclib/phpseclib": "^2.0.11",
@@ -33,8 +33,8 @@
     "psr-4": { "Hyperwallet\\Tests\\" : "tests/Hyperwallet/Tests", "ComposerScript\\" : "src/ComposerScript" }
   },
   "require-dev": {
-    "phpunit/phpunit": "^4.8",
-    "phake/phake": "^2.3",
+    "phpunit/phpunit": "^5.7 || ^9.0",
+    "phake/phake": "^2.3 || ^4.2",
     "satooshi/php-coveralls": "^1.0"
   },
   "scripts": {
