@@ -34,7 +34,7 @@ use Hyperwallet\Response\ErrorResponse;
 use Hyperwallet\Util\ApiClient;
 use Hyperwallet\Util\HyperwalletUUID;
 
-class HyperwalletTest extends \PHPUnit_Framework_TestCase {
+class HyperwalletTest extends \PHPUnit\Framework\TestCase {
 
     public function testConstructor_throwErrorIfUsernameIsEmpty() {
         try {
